pyproject.toml
[tool.poetry]
name = "\\"
version = "0.1.0"
description = "App created with fastapi and redis"
authors = ["SergeiNaum <sergei.naumencko@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = {extras = ["standard"], version = "^0.27.1"}
redis = "^5.0.2"
black = "^24.2.0"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
